メッセージの名前横にカラーアイコンが表示されるように #29
