+++ backend/resources/ts/components/Message.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from 'react'
-import { Box } from '@mui/material'
+import { Box, Avatar, Grid } from '@mui/material'
 import {
   useChatMessageIds,
   useChatMessagesEntities,
@@ -8,6 +8,7 @@ import {
   useParamGroupId,
   useScrollToBottom,
   useUpdateMessages,
+  useStringToColor,
 } from '../app/hooks'
 import EditGroupModal from './EditGroupModal'
 
@@ -39,14 +40,12 @@ const Message = () => {
 
   return (
     <>
-      <Box sx={{ display: 'flex', justifyContent: 'space-around' }}>
-        <h2 className="h2">
-          {groupsEntities[`group${groupId}`]
-            ? groupsEntities[`group${groupId}`].name
-            : ''}
-        </h2>
+      <h2 className="h2">
+        {groupsEntities[`group${groupId}`]
+          ? groupsEntities[`group${groupId}`].name
+          : ''}
         <EditGroupModal groupId={groupId} groupName={groupName} />
-      </Box>
+      </h2>
       <div ref={messageList} className="message">
         <p className="message__note">ここが「{groupName}」の先頭です。</p>
         {chatMessagesIds.map((id: string) =>
@@ -82,14 +81,34 @@ const MessageItem = React.memo(
     // 2個以上の改行を2個改行におさめる
     content = content.replace(/\n{2,}/g, '\n\n')
     const datetime = useFormatDate(created_at)
+    const nameColor = useStringToColor(name)
 
+    function stringToColor(string: string) {
+      let hash = 0
+      let i
+
+      /* eslint-disable no-bitwise */
+      for (i = 0; i < string.length; i += 1) {
+        hash = string.charCodeAt(i) + ((hash << 5) - hash)
+      }
+
+      let color = '#'
+
+      for (i = 0; i < 3; i += 1) {
+        const value = (hash >> (i * 8)) & 0xff
+        color += `00${value.toString(16)}`.substr(-2)
+      }
+      /* eslint-enable no-bitwise */
+
+      return color
+    }
     return (
       <>
         <div className="message__item">
-          <div>
-            <strong className="mr-1">{name}</strong>
-            <small>{datetime}</small>
-          </div>
+          <Grid container>
+            <Avatar title={name} sx={{ bgcolor: nameColor, m: 1 }}></Avatar>
+            <p>{name}</p>
+          </Grid>
           <p>
             {content.split('\n').map((str, index) => (
               <React.Fragment key={index}>
